fix: move types.ts into types folder, delete utils folder
